- Switched QuickFindBar back to using StringFindReplacer since Scintilla's regex matching is too weak for our uses
git-svn-id: https://codelite.svn.sourceforge.net/svnroot/codelite/trunk@2620 9da81c78-c036-0410-9e1f-a2b0375e4b5a

+++ LiteEditor/quickfindbar.h
@@ -36,7 +36,7 @@ class QuickFindBar : public wxPanel
 protected:
     wxTextCtrl  *m_findWhat;
     wxScintilla *m_sci;
-    int          m_flags;
+    size_t       m_flags;
 	wxButton	*m_closeButton;
 
     void DoSearch(bool fwd, bool incr);
